refactor: loadMovies() 메서드 분리
